+++ src/rer/solrpush/interfaces/settings.py
@@ -1,45 +1,42 @@
 # -*- coding: utf-8 -*-
-from jsonschema import validate
-from jsonschema import ValidationError
+from collective.z3cform.datagridfield.registry import DictRow
+from plone.app.vocabularies.catalog import CatalogSource as CatalogSourceBase
+from plone.app.z3cform.widget import RelatedItemsFieldWidget
+from plone.autoform import directives as form
 from plone.supermodel import model
 from rer.solrpush import _
 from zope import schema
-from zope.interface import Invalid
-
-import json
-
-ELEVATE_SCHEMA = {
-    "type": "object",
-    "properties": {
-        "text": {"type": "string"},
-        "id": {"type": "array", "items": {"type": "string"}},
-    },
-}
 
 
-def validate_cfg_json(value):
-    """check that we have at least valid json and its a dict
+class CatalogSource(CatalogSourceBase):
     """
-    try:
-        jv = json.loads(value)
-        if not isinstance(jv, list):
-            raise Invalid(
-                _(
-                    "invalid_json_format",
-                    "JSON is not valid. It should be a list of values.",
-                )
-            )
-        for item in jv:
-            validate(instance=item, schema=ELEVATE_SCHEMA)
-    except (ValueError, ValidationError) as e:
-        raise Invalid(
-            _(
-                "invalid_json",
-                "JSON is not valid, parser complained: ${message}",
-                mapping={"message": e.message},
-            )
-        )
-    return True
+    Collection tile specific catalog source to allow targeted widget.
+    Without this hack, validation doesn't pass
+    """
+
+    def __contains__(self, value):
+        return True  # Always contains to allow lazy handling of removed objs
+
+
+class IElevateRowSchema(model.Schema):
+    text = schema.TextLine(
+        title=_("elevate_row_schema_text_label", default=u"Text"),
+        description=_(
+            "elevate_row_schema_text_help",
+            default=u"The word that should match in the search.",
+        ),
+        required=True,
+    )
+    uid = schema.List(
+        title=_("elevate_row_schema_uid_label", u"Elements"),
+        description=_(
+            "elevate_row_schema_uid_help",
+            u"Select a list of elements to elevate for that search word.",
+        ),
+        value_type=schema.Choice(source=CatalogSource()),
+        required=True,
+    )
+    form.widget("uid", RelatedItemsFieldWidget)
 
 
 class IRerSolrpushConf(model.Schema):
@@ -79,14 +76,14 @@ class IRerSolrpushConf(model.Schema):
         ),
     )
 
-    elevate_schema = schema.SourceText(
-        title=u"Elevate configuration",
-        description=u"Insert a list of values for elevate. Each elevate item "
-        u"should have the following structure: "
-        u'{"text": "some text", "ids": ["12345677"]}.',
+    elevate_schema = schema.List(
+        title=_(u"elevate_schema_label", default=u"Elevate configuration"),
+        description=_(
+            u"elevate_schema_help",
+            default=u"Insert a list of values for elevate.",
+        ),
         required=False,
-        constraint=validate_cfg_json,
-        default=u"[]",
+        value_type=DictRow(title=u"elevate row", schema=IElevateRowSchema),
     )
 
     index_fields = schema.SourceText(