+++ src/rer/solrpush/upgrades.py
@@ -8,3 +8,8 @@ def to_1100(context):
     setup_tool = api.portal.get_tool(name="portal_setup")
     setup_tool.runAllImportStepsFromProfile("profile-rer.solrpush:to_1100")
     setup_tool.runImportStepFromProfile(default_profile, "plone.app.registry")
+
+def to_1200(context):
+    setup_tool = api.portal.get_tool(name="portal_setup")
+    setup_tool.runAllImportStepsFromProfile("profile-rer.solrpush:1200")
+    setup_tool.runImportStepFromProfile(default_profile, "plone.app.registry")